requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
click>=8.1.0
pydantic>=2.0.0
rich>=13.0.0
jinja2>=3.1.0
pandas>=2.0.0
pytest>=7.4.0
pytest-mock>=3.12.0
